edited freeswitch

+++ libs/sofia-sip/libsofia-sip-ua/sdp/sdp.c
@@ -1469,7 +1469,7 @@ int sdp_media_cmp(sdp_media_t const *a, sdp_media_t const *b)
 
   if (a->m_proto != b->m_proto)
     return a->m_proto < b->m_proto ? -1 : 1;
-  if (a->m_proto == sdp_media_x)
+  if ((int)a->m_proto == (int)sdp_media_x)
     if ((rv = su_strcmp(a->m_proto_name, b->m_proto_name)))
       return rv;
